Revert "Moved just icons.js to shared folder"
