+++ packages/vue/src/assets/icons.js
