+++ packages/vue/src/components/atoms/SfIcon/SfIcon.js
@@ -1,4 +1,4 @@
-import { icons } from "@storefrontui/shared/icons/icons.js";
+import { icons } from "../../../assets/icons";
 
 const HEX_REGEX = /^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$/;
 const RGB_REGEX = /rgb\((\d{1,3}), (\d{1,3}), (\d{1,3})\)/;
